+++ ethereum/executionclient/src/integration-test/java/tech/pegasys/teku/ethereum/executionclient/rest/RestBuilderClientTest.java
@@ -15,6 +15,10 @@ package tech.pegasys.teku.ethereum.executionclient.rest;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.spy;
 import static tech.pegasys.teku.spec.SpecMilestone.BELLATRIX;
 import static tech.pegasys.teku.spec.SpecMilestone.CAPELLA;
 import static tech.pegasys.teku.spec.SpecMilestone.DENEB;
@@ -26,11 +30,13 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.io.Resources;
 import java.io.IOException;
 import java.io.UncheckedIOException;
+import java.net.ConnectException;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.Locale;
 import java.util.Optional;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeoutException;
 import java.util.function.Consumer;
 import java.util.regex.Pattern;
 import okhttp3.OkHttpClient;
@@ -45,13 +51,16 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.TestTemplate;
 import tech.pegasys.teku.bls.BLSPublicKey;
+import tech.pegasys.teku.ethereum.executionclient.BuilderApiMethod;
 import tech.pegasys.teku.ethereum.executionclient.schema.BuilderApiResponse;
+import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.json.JsonUtil;
 import tech.pegasys.teku.infrastructure.json.exceptions.MissingRequiredFieldException;
 import tech.pegasys.teku.infrastructure.json.types.DeserializableTypeDefinition;
 import tech.pegasys.teku.infrastructure.ssz.SszData;
 import tech.pegasys.teku.infrastructure.ssz.SszList;
 import tech.pegasys.teku.infrastructure.ssz.sos.SszDeserializeException;
+import tech.pegasys.teku.infrastructure.time.StubTimeProvider;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.SpecMilestone;
@@ -98,11 +107,13 @@ class RestBuilderClientTest {
   private final OkHttpClient okHttpClient = new OkHttpClient.Builder().build();
   private final MockWebServer mockWebServer = new MockWebServer();
 
+  private StubTimeProvider timeProvider;
   private Spec spec;
   private SpecMilestone milestone;
   private OkHttpRestClient okHttpRestClient;
   private SchemaDefinitionsBellatrix schemaDefinitions;
 
+  private RestBuilderClientOptions restBuilderClientOptions;
   private RestBuilderClient restBuilderClient;
 
   private String signedValidatorRegistrationsRequest;
@@ -117,6 +128,7 @@ class RestBuilderClientTest {
   @BeforeEach
   void setUp(final SpecContext specContext) throws IOException {
     mockWebServer.start();
+    timeProvider = StubTimeProvider.withTimeInMillis(UInt64.ONE);
 
     spec = specContext.getSpec();
     final String endpoint = "http://localhost:" + mockWebServer.getPort();
@@ -152,7 +164,9 @@ class RestBuilderClientTest {
                 "data")
             .orElseThrow();
 
-    restBuilderClient = new RestBuilderClient(okHttpRestClient, spec, true);
+    restBuilderClientOptions = RestBuilderClientOptions.DEFAULT;
+    restBuilderClient =
+        new RestBuilderClient(restBuilderClientOptions, okHttpRestClient, timeProvider, spec, true);
   }
 
   @AfterEach
@@ -207,7 +221,7 @@ class RestBuilderClientTest {
               assertThat(response.payload()).isNull();
             });
 
-    verifyPostRequestJson("/eth/v1/builder/validators", signedValidatorRegistrationsRequest);
+    verifyPostRequestSsz("/eth/v1/builder/validators", signedValidatorRegistrations);
   }
 
   @TestTemplate
@@ -232,6 +246,8 @@ class RestBuilderClientTest {
 
     final String unknownValidatorError = "{\"code\":400,\"message\":\"unknown validator\"}";
 
+    // Both ssz and json requests will fail
+    mockWebServer.enqueue(new MockResponse().setResponseCode(400).setBody(unknownValidatorError));
     mockWebServer.enqueue(new MockResponse().setResponseCode(400).setBody(unknownValidatorError));
 
     final SszList<SignedValidatorRegistration> signedValidatorRegistrations =
@@ -245,8 +261,10 @@ class RestBuilderClientTest {
               assertThat(response.errorMessage()).isEqualTo(unknownValidatorError);
             });
 
+    verifyPostRequestSsz("/eth/v1/builder/validators", signedValidatorRegistrations);
     verifyPostRequestJson("/eth/v1/builder/validators", signedValidatorRegistrationsRequest);
 
+    // Only json will fail (ssz backoff enabled)
     mockWebServer.enqueue(
         new MockResponse().setResponseCode(500).setBody(INTERNAL_SERVER_ERROR_MESSAGE));
 
@@ -262,9 +280,104 @@ class RestBuilderClientTest {
   }
 
   @TestTemplate
+  void registerValidators_retryWithJsonAfterSszFailure() {
+    final String errorMsg = "{\"code\":415,\"message\":\"unsupported media type\"}";
+    mockWebServer.enqueue(new MockResponse().setResponseCode(415).setBody(errorMsg));
+    mockWebServer.enqueue(
+        new MockResponse().setResponseCode(200).setBody(signedBuilderBidResponseJson));
+
+    final SszList<SignedValidatorRegistration> signedValidatorRegistrations =
+        createSignedValidatorRegistrations();
+
+    assertThat(restBuilderClient.registerValidators(SLOT, signedValidatorRegistrations))
+        .succeedsWithin(WAIT_FOR_CALL_COMPLETION)
+        .satisfies(response -> assertThat(response.isSuccess()).isTrue());
+
+    verifyPostRequestSsz("/eth/v1/builder/validators", signedValidatorRegistrations);
+    verifyPostRequestJson("/eth/v1/builder/validators", signedValidatorRegistrationsRequest);
+
+    assertThat(mockWebServer.getRequestCount()).isEqualTo(2);
+  }
+
+  @TestTemplate
+  void registerValidators_trySszAfterBackoffTime() {
+    final String errorMsg = "{\"code\":400,\"message\":\"bad request\"}";
+    mockWebServer.enqueue(new MockResponse().setResponseCode(400).setBody(errorMsg));
+    mockWebServer.enqueue(
+        new MockResponse().setResponseCode(200).setBody(signedBuilderBidResponseJson));
+    mockWebServer.enqueue(
+        new MockResponse().setResponseCode(200).setBody(signedBuilderBidResponseJson));
+
+    final SszList<SignedValidatorRegistration> signedValidatorRegistrations =
+        createSignedValidatorRegistrations();
+
+    // First call, will try SSZ and fallback into JSON
+    assertThat(restBuilderClient.registerValidators(SLOT, signedValidatorRegistrations))
+        .succeedsWithin(WAIT_FOR_CALL_COMPLETION)
+        .satisfies(response -> assertThat(response.isSuccess()).isTrue());
+
+    verifyPostRequestSsz("/eth/v1/builder/validators", signedValidatorRegistrations);
+    verifyPostRequestJson("/eth/v1/builder/validators", signedValidatorRegistrationsRequest);
+
+    // After backoff period, will try SSZ again
+    timeProvider.advanceTimeBy(Duration.ofDays(1));
+    assertThat(restBuilderClient.registerValidators(SLOT, signedValidatorRegistrations))
+        .succeedsWithin(WAIT_FOR_CALL_COMPLETION)
+        .satisfies(response -> assertThat(response.isSuccess()).isTrue());
+
+    verifyPostRequestSsz("/eth/v1/builder/validators", signedValidatorRegistrations);
+
+    assertThat(mockWebServer.getRequestCount()).isEqualTo(3);
+  }
+
+  @TestTemplate
+  void registerValidators_shouldNotFallbackWhenTimingOut() {
+    // Creates a RestBuilderClient that always timeout
+    restBuilderClient =
+        new RestBuilderClient(
+            forcedTimeoutRestBuilderClientOptions(), okHttpRestClient, timeProvider, spec, true);
+
+    final SszList<SignedValidatorRegistration> signedValidatorRegistrations =
+        createSignedValidatorRegistrations();
+
+    assertThat(restBuilderClient.registerValidators(SLOT, signedValidatorRegistrations))
+        .failsWithin(WAIT_FOR_CALL_COMPLETION)
+        .withThrowableThat()
+        .withCauseInstanceOf(TimeoutException.class);
+
+    verifyPostRequestSsz("/eth/v1/builder/validators", signedValidatorRegistrations);
+    // Check that we do not fallback into JSON (only 1 request)
+    assertThat(mockWebServer.getRequestCount()).isEqualTo(1);
+  }
+
+  @TestTemplate
+  void registerValidators_shouldNotFallbackWhenFailingForNonHttpReasons() {
+    okHttpRestClient = spy(okHttpRestClient);
+    // Mock asyncPost with ssz to fail
+    doReturn(SafeFuture.failedFuture(new ConnectException()))
+        .when(okHttpRestClient)
+        .postAsync(eq(BuilderApiMethod.REGISTER_VALIDATOR.getPath()), any(), eq(true));
+    restBuilderClient =
+        new RestBuilderClient(restBuilderClientOptions, okHttpRestClient, timeProvider, spec, true);
+
+    final SszList<SignedValidatorRegistration> signedValidatorRegistrations =
+        createSignedValidatorRegistrations();
+
+    assertThat(restBuilderClient.registerValidators(SLOT, signedValidatorRegistrations))
+        .failsWithin(WAIT_FOR_CALL_COMPLETION)
+        .withThrowableThat()
+        .withCauseInstanceOf(ConnectException.class);
+
+    // No requests are made (including no fallback to json)
+    assertThat(mockWebServer.getRequestCount()).isEqualTo(0);
+  }
+
+  @TestTemplate
   void getHeader_success_doesNotSetUserAgentHeader() {
 
-    restBuilderClient = new RestBuilderClient(okHttpRestClient, spec, false);
+    restBuilderClient =
+        new RestBuilderClient(
+            restBuilderClientOptions, okHttpRestClient, timeProvider, spec, false);
 
     mockWebServer.enqueue(
         new MockResponse().setResponseCode(200).setBody(signedBuilderBidResponseJson));
@@ -622,6 +735,10 @@ class RestBuilderClientTest {
     verifyRequest("POST", apiPath, Optional.of(requestBody), Optional.empty());
   }
 
+  private <T extends SszData> void verifyPostRequestSsz(final String apiPath, final T requestBody) {
+    verifyRequest("POST", apiPath, Optional.empty(), Optional.of(requestBody));
+  }
+
   private void verifyRequest(
       final String method,
       final String apiPath,
@@ -724,5 +841,10 @@ class RestBuilderClientTest {
   private static String changeResponseVersion(final String json, final SpecMilestone newVersion) {
     return json.replaceFirst(
         "(?<=version\":\\s?\")\\w+", newVersion.toString().toLowerCase(Locale.ROOT));
+  }
+
+  private RestBuilderClientOptions forcedTimeoutRestBuilderClientOptions() {
+    return new RestBuilderClientOptions(
+        Duration.ofSeconds(0), Duration.ofSeconds(0), Duration.ofSeconds(0), Duration.ofSeconds(0));
   }
 }